requirements.txt
librosa==0.6.1
matplotlib==2.2.2
numpy==1.15.4
opencv-python==3.4.1.15
Pillow==5.1.0
pydub==0.22.1
PyYAML==5.1.1
scikit-image==0.14.0
scikit-learn==0.20.0
scipy==1.1.0
sklearn==0.0
tensorboard==1.9.0
tensorboardX==1.2
torch==0.4.1
torchvision==0.2.1
wget==3.2
tqdm==4.23.4